+++ hutool-core/src/main/java/cn/hutool/core/util/ArrayUtil.java
@@ -1651,6 +1651,26 @@ public class ArrayUtil extends PrimitiveArrayUtil {
 	}
 
 	/**
+	 * 按照指定规则，将一种类型的数组转换为另一种类型
+	 *
+	 * @param array               被转换的数组
+	 * @param targetComponentType 目标的元素类型
+	 * @param func                转换规则函数
+	 * @param <T>                 原数组类型
+	 * @param <R>                 目标数组类型
+	 * @return 转换后的数组
+	 * @since 5.5.8
+	 */
+	public static <T, R> R[] map(Object array, Class<R> targetComponentType, Function<? super T, ? extends R> func) {
+		final int length = length(array);
+		final R[] result = newArray(targetComponentType, length);
+		for (int i = 0; i < length; i++) {
+			result[i] = func.apply(get(array, i));
+		}
+		return result;
+	}
+
+	/**
 	 * 按照指定规则，将一种类型的数组元素提取后转换为List
 	 *
 	 * @param array 被转换的数组